Refactor test units & utilities
