+++ app/tests/support/mocks/servers.py
@@ -1,10 +1,12 @@
 import asyncio as io
 
 from ....utils.middlewares import accept_conn
+from ....utils.emitter import Responder
 from ....settings import HOST_URL, HOST_PORT
 
 class MockServerSpec:
     _server: io.base_events.Server
+    _pulse: io.Task
 
     def __init__(self, host: str, port: int):
         self._host = host
@@ -19,13 +21,23 @@ class MockServerSpec:
     def is_closed(self):
         return self._server.is_serving()
 
-    async def terminate(self):
+    async def stop_pulse(self):
         try:
-            self._server.close()
+            self._pulse.cancel()
+            await self._pulse
         except:
             ...
-        finally:
+
+    async def stop_server(self):
+        try:
+            self._server.close()
             await self._server.wait_closed()
+        except:
+            ...
+
+    async def terminate(self):
+        await self.stop_pulse()
+        await self.stop_server()
 
 class AppServerSpec(MockServerSpec):
     def __init__(self, host=HOST_URL, port=HOST_PORT):
@@ -35,6 +47,7 @@ class AppServerSpec(MockServerSpec):
         self._server = await io.start_server(
             accept_conn, host=self._host, port=self._port
         )
+        self._pulse = io.create_task(Responder.pulse())
         return self
 
 class ConnectivityMockServer(MockServerSpec):
